feat(trading): eligibility active rewards (#6010)
