+++ apps/trading/components/competitions/games-container.tsx
@@ -1,6 +1,9 @@
-import { ActiveRewardCard } from '../rewards-container/active-rewards';
 import { useT } from '../../lib/use-t';
 import { type EnrichedRewardTransfer } from '../../lib/hooks/use-rewards';
+import { useVegaWallet } from '@vegaprotocol/wallet-react';
+import { useStakeAvailable } from '../../lib/hooks/use-stake-available';
+import { useMyTeam } from '../../lib/hooks/use-my-team';
+import { ActiveRewardCard } from '../rewards-container/reward-card';
 
 export const GamesContainer = ({
   data,
@@ -10,6 +13,19 @@ export const GamesContainer = ({
   currentEpoch: number;
 }) => {
   const t = useT();
+  const { pubKey } = useVegaWallet();
+  const { team } = useMyTeam();
+  const { stakeAvailable, isEligible, requiredStake } = useStakeAvailable();
+
+  const requirements = pubKey
+    ? {
+        isEligible,
+        stakeAvailable,
+        requiredStake,
+        team,
+        pubKey,
+      }
+    : undefined;
 
   if (!data || data.length === 0) {
     return (
@@ -35,6 +51,7 @@ export const GamesContainer = ({
             key={i}
             transferNode={game}
             currentEpoch={currentEpoch}
+            requirements={requirements}
           />
         );
       })}
