+++ apps/trading/lib/hooks/Rewards.graphql
@@ -111,6 +111,22 @@ query ActiveRewards(
   }
 }
 
+query TWAP($partyId: ID!, $assetId: ID!, $gameId: ID!, $epochSeq: Int) {
+  timeWeightedNotionalPosition(
+    partyId: $partyId
+    gameId: $gameId
+    assetId: $assetId
+    epochSeq: $epochSeq
+  ) {
+    assetId
+    partyId
+    gameId
+    timeWeightedNotionalPosition
+    lastUpdated
+    epoch
+  }
+}
+
 query RewardsHistory(
   $partyId: ID!
   $epochRewardSummariesPagination: Pagination
